GH-14: Split core into multiple files
